Divide everything into v1, v2, and common (unexported).
Did some refactoring, like putting `firebaseConfig` into common.
This means we have absolutely no dependencies on v1 in common or
v2. Some files are left at the root because they were modified
in other CLs that are outstanding and I want to linearize the
changes to minimize conflicts.

While working, I removed some legacy workarounds that should no
longer apply. For example, we no longer need to load fireaseConfig
from runtime config because all supported versions of the CLI
set the FIREBASE_CONFIG environment variable. This meant we
could remove all of setup.ts because all monkeypatches were for
outdated runtimes. Since there is no longer an environment
change between importing v1/index and a subpackage, I've gone
ahead and exported the providers for v1 to minimize customer
frustration with the v1 namespace.
