+++ integration_test/firebase.json
@@ -5,5 +5,9 @@
   "firestore": {
     "rules": "firestore.rules",
     "indexes": "firestore.indexes.json"
+  },
+  "functions": {
+    "source": "functions",
+    "codebase": "integration-tests"
   }
 }
