+++ spec/fixtures/sources/commonjs-grouped/g1.js
@@ -1,4 +1,4 @@
-const functions = require("../../../../src/index");
+const functions = require("../../../../src/v1");
 
 exports.groupedhttp = functions.https.onRequest((req, resp) => {
     resp.status(200).send("PASS");
