+++ spec/v1/config.spec.ts
@@ -25,7 +25,7 @@ import * as fs from 'fs';
 import * as process from 'process';
 import Sinon = require('sinon');
 
-import * as config from '../../src/config';
+import { config, resetCache } from '../../src/v1/config';
 
 describe('config()', () => {
   let readFileSync: Sinon.SinonStub;
@@ -43,8 +43,7 @@ describe('config()', () => {
   });
 
   afterEach(() => {
-    delete config.config.singleton;
-    (config as any).firebaseConfigCache = null;
+    resetCache();
     delete process.env.FIREBASE_CONFIG;
     delete process.env.CLOUD_RUNTIME_CONFIG;
     delete process.env.K_CONFIGURATION;
@@ -60,10 +59,7 @@ describe('config()', () => {
       .returns(Buffer.from(json));
 
     process.env.K_CONFIGURATION = 'my-service';
-    expect(() => config.config()).to.throw(
-      Error,
-      /transition to using environment variables/
-    );
+    expect(config).to.throw(Error, /transition to using environment variables/);
   });
 
   it('loads config values from .runtimeconfig.json', () => {
@@ -74,66 +70,8 @@ describe('config()', () => {
     readFileSync
       .withArgs('/srv/.runtimeconfig.json')
       .returns(Buffer.from(json));
-    const loaded = config.config();
+    const loaded = config();
     expect(loaded).to.not.have.property('firebase');
     expect(loaded).to.have.property('foo', 'bar');
-  });
-
-  it('does not provide firebase config if .runtimeconfig.json not invalid', () => {
-    readFileSync.withArgs('/srv/.runtimeconfig.json').returns('invalid JSON');
-    expect(config.firebaseConfig()).to.be.null;
-  });
-
-  it('does not provide firebase config if .ruuntimeconfig.json has no firebase property', () => {
-    readFileSync
-      .withArgs('/srv/.runtimeconfig.json')
-      .returns(Buffer.from('{}'));
-    expect(config.firebaseConfig()).to.be.null;
-  });
-
-  it('loads Firebase configs from FIREBASE_CONFIG env variable', () => {
-    process.env.FIREBASE_CONFIG = JSON.stringify({
-      databaseURL: 'foo@firebaseio.com',
-    });
-    expect(config.firebaseConfig()).to.have.property(
-      'databaseURL',
-      'foo@firebaseio.com'
-    );
-  });
-
-  it('loads Firebase configs from FIREBASE_CONFIG env variable pointing to a file', () => {
-    const oldEnv = process.env;
-    (process as any).env = {
-      ...oldEnv,
-      FIREBASE_CONFIG: '.firebaseconfig.json',
-    };
-    try {
-      readFileSync.returns(
-        Buffer.from('{"databaseURL": "foo@firebaseio.com"}')
-      );
-      expect(config.firebaseConfig()).to.have.property(
-        'databaseURL',
-        'foo@firebaseio.com'
-      );
-    } finally {
-      (process as any).env = oldEnv;
-    }
-  });
-
-  it('accepts alternative locations for config file', () => {
-    process.env.CLOUD_RUNTIME_CONFIG = 'another.json';
-    const json = JSON.stringify({ foo: 'bar', firebase: {} });
-    readFileSync.withArgs('another.json').returns(Buffer.from(json));
-    expect(config.firebaseConfig()).to.not.be.null;
-    expect(config.config()).to.have.property('foo', 'bar');
-  });
-
-  it('accepts full JSON in env.CLOUD_RUNTIME_CONFIG', () => {
-    process.env.CLOUD_RUNTIME_CONFIG = JSON.stringify({
-      foo: 'bar',
-      firebase: {},
-    });
-    expect(config.firebaseConfig()).to.not.be.null;
-    expect(config.config()).to.have.property('foo', 'bar');
   });
 });
