+++ spec/v1/providers/pubsub.spec.ts
@@ -21,9 +21,9 @@
 // SOFTWARE.
 
 import { expect } from 'chai';
-import { Event } from '../../../src/index';
-import * as functions from '../../../src/index';
-import * as pubsub from '../../../src/providers/pubsub';
+import { Event } from '../../../src/v1';
+import * as functions from '../../../src/v1';
+import * as pubsub from '../../../src/v1/providers/pubsub';
 
 describe('Pubsub Functions', () => {
   describe('pubsub.Message', () => {
