+++ src/utilities/encoder.ts
