+++ src/v1/cloud-functions.ts
@@ -22,13 +22,13 @@
 
 import { Request, Response } from 'express';
 import * as _ from 'lodash';
+import { warn } from '../logger';
 import {
   DEFAULT_FAILURE_POLICY,
   DeploymentOptions,
   FailurePolicy,
   Schedule,
 } from './function-configuration';
-import { warn } from './logger';
 export { Request, Response };
 import {
   convertIfPresent,
@@ -36,8 +36,8 @@ import {
   Duration,
   durationFromSeconds,
   serviceAccountFromShorthand,
-} from './common/encoding';
-import { ManifestEndpoint, ManifestRequiredAPI } from './runtime/manifest';
+} from '../common/encoding';
+import { ManifestEndpoint, ManifestRequiredAPI } from '../runtime/manifest';
 
 /** @hidden */
 const WILDCARD_REGEX = new RegExp('{[^/{}]*}', 'g');
