+++ src/v1/index.ts
@@ -32,10 +32,9 @@ import * as storage from './providers/storage';
 import * as tasks from './providers/tasks';
 import * as testLab from './providers/testLab';
 
-import * as apps from './apps';
+import * as apps from '../apps';
+import * as logger from '../logger';
 import { handler } from './handler-builder';
-import * as logger from './logger';
-import { setup } from './setup';
 
 const app = apps.apps();
 
@@ -60,5 +59,3 @@ export * from './cloud-functions';
 export * from './config';
 export * from './function-builder';
 export * from './function-configuration';
-
-setup();
