+++ src/v1/providers/storage.ts
@@ -20,12 +20,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+import { firebaseConfig } from '../../common/config';
 import {
   CloudFunction,
   EventContext,
   makeCloudFunction,
 } from '../cloud-functions';
-import { firebaseConfig } from '../config';
 import { DeploymentOptions } from '../function-configuration';
 
 /** @hidden */
