+++ tsconfig.release.json
@@ -12,7 +12,7 @@
   },
   "files": [
     "./src/types/global.d.ts",
-    "./src/index.ts",
+    "./src/v1/index.ts",
     "./src/logger/index.ts",
     "./src/logger/compat.ts",
     "./src/v2/index.ts",
